requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
streamlit>=1.28.0
requests>=2.31.0

# Azure dependencies
azure-core>=1.29.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# Database
SQLAlchemy>=2.0.0
pyodbc>=5.0.0

# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# AI/ML frameworks
crewai>=0.1.0
langchain-openai>=0.0.5

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
